dev: update read
