--- examples/utils.py
@@ -1,12 +1,0 @@
-import os
-import random
-
-import numpy as np
-import tensorflow as tf
-
-
-def set_seed(seed):
-    random.seed(seed)
-    np.random.seed(seed)
-    os.environ["PYTHONHASHSEED"] = str(seed)
-    tf.random.set_seed(seed)
